+++ CHANGELOG.md
@@ -2,6 +2,6 @@
 
 ## v0.3.0
 
-* (C) Extracted from Tideland Go Library as part of split
-* (C) Reimplemented stopwatch
-
+- (C) Extracted from Tideland Go Library as part of split
+- (C) Reimplemented stopwatch
+- (C) Reimplemented stay-set indicator